requirements.txt
# Core dependencies
apache-airflow==2.7.1
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
joblib>=1.2.0

# MLflow integration
mlflow>=2.8.0
boto3>=1.26.0  # For S3 artifact storage with MLflow

# Database connectors
psycopg2-binary>=2.9.5  # PostgreSQL connector
pymysql>=1.0.2  # MySQL connector

# Web application
streamlit>=1.22.0
requests>=2.28.0

# Optional: Uncomment if needed
# xgboost>=1.7.0
# lightgbm>=3.3.0
# tensorflow>=2.12.0
# torch>=2.0.0

# Development tools
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
